+++ tests/testthat/helper-run-data-type-tests.R
@@ -7,6 +7,36 @@ runDataTypeTests <- function(db, redshift=FALSE) {
     expect_equal(res$counter, events$counter)
   })
 
+  test_that("bigint works", {
+    events <- data.frame(bigcounter=c(1, 2))
+    dbxInsert(db, "events", events)
+
+    res <- dbxSelect(db, "SELECT * FROM events ORDER BY id")
+    expect_equal(res$bigcounter, events$bigcounter)
+  })
+
+  test_that("large bigint", {
+    dbxExecute(db, "INSERT INTO events (bigcounter) VALUES (9007199254740991)")
+    dbxSelect(db, "SELECT * FROM events ORDER BY id")
+    dbxExecute(db, "INSERT INTO events (bigcounter) VALUES (9007199254740992)")
+
+    # TODO change behavior
+    # expect_error(dbxSelect(db, "SELECT * FROM events ORDER BY id"), "bigint value outside range of numeric")
+    res <- dbxSelect(db, "SELECT * FROM events ORDER BY id")
+    expect_equal(res$bigcounter, c(9007199254740991, 9007199254740992))
+  })
+
+  test_that("small bigint", {
+    dbxExecute(db, "INSERT INTO events (bigcounter) VALUES (-9007199254740991)")
+    dbxSelect(db, "SELECT * FROM events ORDER BY id")
+    dbxExecute(db, "INSERT INTO events (bigcounter) VALUES (-9007199254740992)")
+
+    # TODO change behavior
+    # expect_error(dbxSelect(db, "SELECT * FROM events ORDER BY id"), "bigint value outside range of numeric")
+    res <- dbxSelect(db, "SELECT * FROM events ORDER BY id")
+    expect_equal(res$bigcounter, c(-9007199254740991, -9007199254740992))
+  })
+
   test_that("float works", {
     events <- data.frame(speed=c(1.2, 3.4))
     dbxInsert(db, "events", events)
@@ -258,12 +288,41 @@ runDataTypeTests <- function(db, redshift=FALSE) {
     expect_equal(blob::as.blob(res$image), events$image)
   })
 
+  test_that("empty blob works", {
+    events <- data.frame(city=c("Boston"))
+    dbxInsert(db, "events", events)
+
+    sql <- "SELECT id, image FROM events WHERE image IS NULL"
+    res <- dbxSelect(db, sql)
+    expect_equal(nrow(res), 1)
+
+    if (isRMySQL(db)) {
+      # RMySQL cannot return binary data
+      expect_equal(res$image[[1]], as.character(NA))
+    } else {
+      expect_null(res$image[[1]])
+
+      dbxUpdate(db, "events", res, where_cols=c("id"))
+      res <- dbxSelect(db, sql)
+      expect_equal(nrow(res), 1)
+    }
+  })
+
   test_that("ts uses observation values", {
     events <- data.frame(counter=ts(1:3, start=c(2018, 1), end=c(2018, 3), frequency=12))
     dbxInsert(db, "events", events)
 
     res <- dbxSelect(db, "SELECT * FROM events ORDER BY id")
     expect_equal(res$counter, as.integer(events$counter))
+  })
+
+  # TODO raise error?
+  test_that("difftime works", {
+    events <- data.frame(city=as.difftime(c("12:30:55", "16:59:59")))
+    dbxInsert(db, "events", events)
+
+    res <- dbxSelect(db, "SELECT * FROM events")
+    expect_equal(res$city, as.character(events$city))
   })
 
   # very important