+++ NEWS.md
@@ -1,3 +1,9 @@
+# dbx 0.2.4
+
+- Added `dbxExecute` function
+- Added support for complex types
+- Fixed issues with NULL blobs with Postgres and SQLite
+
 # dbx 0.2.3
 
 - Added parameters for `dbxSelect`